Completed Initial hashing setup

+++ src/utils/hashTable.h
@@ -1,18 +1,19 @@
 #ifndef HASH_TABLE_H
 #define HASH_TABLE_H
 
-#include "uthash.h"
+#include "../../lib/uthash.h"
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 
 struct hash_entry {
-  char key[256];       // key data (255 bytes + null terminator)
-  uint32_t key_len;    // actual key length
-  UT_hash_handle hh;   // makes structure hashable
+  char key[256];     // key data (255 bytes + null terminator)
+  uint32_t key_len;  // actual key length
+  UT_hash_handle hh; // makes structure hashable
 };
 
-// Creates an empty hash table, returns null as initial table as UThash automatically handles table creation with first operation
+// Creates an empty hash table, returns null as initial table as UThash
+// automatically handles table creation with first operation
 static inline struct hash_entry *create_table(void) { return NULL; }
 
 // Adds a key to the hash table
@@ -27,8 +28,8 @@ static inline void add_key(struct hash_entry **table, const void *key,
 }
 
 // Checks if a key exists in the table
-static inline int key_in_table(struct hash_entry **table, const void *key,
-                               uint32_t key_len) {
+static inline uint8_t key_in_table(struct hash_entry **table, const void *key,
+                                   uint32_t key_len) {
   struct hash_entry *entry;
 
   HASH_FIND(hh, *table, key, key_len, entry);
@@ -59,7 +60,6 @@ static inline void free_table(struct hash_entry **table) {
     HASH_DEL(*table, current);
     free(current);
   }
-  
 }
 
 #endif // HASH_TABLE_H
